+++ code/src/components/CheckBox.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-const CheckBoxes = (props) => {
+const CheckBox = (props) => {
     const {nextQuestion, setPersonality, personality, setPerson, person } = props
         return (
             <section className='wrap'>
@@ -9,33 +9,33 @@ const CheckBoxes = (props) => {
                     <p>What personality applies to you?</p>
                     <form className="form-checkbox">
                         <label htmlFor='personality'>
-                            fearless
                             <input
-                                className='input-class'
+                                className='input-space'
                                 type='checkbox'
                                 checked={personality}
                                 onChange={setPersonality}
                                 id='personality'
                                 name='personality'
                             />
+                            fearless
                         </label>
                         <label htmlFor='person'>
-                            judgemental
                             <input
-                                className='input-class'
+                                className='input-space'
                                 type='checkbox'
                                 checked={person}
                                 onChange={setPerson}
                                 id='person'
                                 name='person'
                             />
+                            judgemental
                         </label>
                     </form>  
-                    <button className='btn' onClick={nextQuestion}>Go to next question</button>
+                    <button className='btn' onClick={nextQuestion}>Next question</button>
                 </div>
             </section>
         )
 }
 
-export default CheckBoxes; 
+export default CheckBox;
